+++ src/main/scala/com/rayrobdod/json/union/JsonValue.scala
@@ -124,6 +124,7 @@ object JsonValue {
 	 * Convert a CborValue into a JsonValue, where ByteStrs are instead converted into
 	 * hexencoded strings.
 	 */
+	@deprecated("doesn't handle incompatible numbers; use cborValue2JsonValueEither then fold that method's return value", "3.1")
 	def cborValueHexencodeByteStr(x:CborValue):JsonValue = x match {
 		case CborValue.CborValueString(s) => JsonValueString(s)
 		case CborValue.CborValueBoolean(b) => JsonValueBoolean(b)
@@ -132,6 +133,17 @@ object JsonValue {
 			s.flatMap{byte => ("00" + (0xFF & byte.intValue).toHexString).takeRight(2)}
 		))
 		case CborValue.CborValueNull => JsonValueNull
+	}
+	
+	/** 
+	 * Convert a CborValue into a JsonValue if there is an equivalent JsonValue; else return a UnsuccessfulTypeCoersion.
+	 */
+	def cborValue2JsonValueEither(x:CborValue):Either[Either[Array[Byte], CborValue.Rational], JsonValue] = x match {
+		case CborValue.CborValueString(s) => Right(JsonValueString(s))
+		case CborValue.CborValueBoolean(b) => Right(JsonValueBoolean(b))
+		case CborValue.CborValueNumber(r) => r.tryToBigDecimal.fold[Either[Either[Array[Byte], CborValue.Rational], JsonValue]](Left(Right(r))){x => Right(JsonValueNumber(x))}
+		case CborValue.CborValueByteStr(s) => Left(Left(s))
+		case CborValue.CborValueNull => Right(JsonValueNull)
 	}
 	
 	private class ReturnLeft(msg:String) extends Function1[Any, Either[(String, Int), Nothing]] with Function0[Either[(String, Int), Nothing]] {